+++ packages/react-native-editor/CHANGELOG.md
@@ -10,6 +10,7 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+-   [**] Video block: Fix logic for displaying empty state based on source presence [#58015]
 
 ## 1.111.0
 -   [**] Image block media uploads display a custom error message when there is no internet connection [#56937]